requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
pandas>=2.0.0
ultralytics>=8.0.0
scikit-learn>=1.3.0
scipy>=1.11.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
streamlit>=1.28.0
pyyaml>=6.0
deep-sort-realtime>=1.3.2